Initial Commit
CypherCrack uses basic cypher (caesar) techniques to attempt to break a cypher.  This was originally intended to be used to solve the "Daily Cryptoquote" in the newspaper.  GraphHist uses matplotlib to plot the histogram generated by CypherCrack to give the user a visual representation of the letter usage in the cypher.